Refactor DataFetcher component and replace AirtableComponent; implement custom hook for scan handling
